refactor: 1차 피드백 반영
- SelectorType 수정
- types 파일명 변경
