+++ src/js/templates/manageItemTemplate.ts
@@ -1,4 +1,4 @@
-import { ItemType } from '../types/types';
+import { ItemType } from '../types';
 
 export const sectionTemplate = {
   inputContainer(): string {
