+++ src/js/validates/validates.ts
@@ -1,4 +1,4 @@
-import { ItemType } from '../types/types';
+import { ItemType } from '../types';
 import { ERROR_MESSAGE, ITEM, MONEY } from '../constants/constants';
 
 export const validateAddItemInput = ({ name, price, quantity }: ItemType) => {
